+++ indico/web/client/js/custom_elements/ind_date_picker.js
@@ -7,6 +7,7 @@
 
 import {formatDate} from 'indico/utils/date_format';
 import {createDateParser} from 'indico/utils/date_parser';
+import {topBottomPosition} from 'indico/utils/positioning';
 
 import './ind_date_picker.scss';
 
@@ -92,15 +93,10 @@ customElements.define(
       });
 
       function openDialog() {
-        indCalendar.toggleAttribute('data-position-check');
-        requestAnimationFrame(() => {
-          const inputRect = input.getBoundingClientRect();
-          const dialogRect = indCalendar.firstElementChild.getBoundingClientRect();
-          const dialogBottom = inputRect.bottom + dialogRect.height;
-          const dialogWillFitBelow = dialogBottom < window.innerHeight;
-          indCalendar.style.setProperty('--dialog-bottom', dialogWillFitBelow ? 'auto' : '100%');
-          indCalendar.toggleAttribute('data-position-check');
-          indCalendar.open = true;
+        topBottomPosition(indCalendar.querySelector('dialog'), input, {
+          expand() {
+            indCalendar.open = true;
+          },
         });
       }
     }