+++ CHANGES.rst
@@ -76,6 +76,7 @@ Internal Changes
 - Add ``<ind-date-picker>`` custom element (:pr:`6371`, thanks :user:`foxbunny`)
 - Use native ESM for webpack config files (:pr:`6389`)
 - Add ``<ind-combobox>`` custom element (:pr:`6310`, thanks :user:`foxbunny`)
+- Add ``<ind-combo-box>`` custom element (:pr:`6310`, thanks :user:`foxbunny`)
 
 
 Version 3.3.2